final blockout
