--- Assets/ik hou niet zo van sorteren.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 725a342b225c9f4499f368134042aff7
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
