--- Assets/ik hou niet zo van sorteren/TeleportTrigger.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 1056d8df3b68b8b4ba6d12635d737f15
